+++ FE/src/pages/login/LoginOk.jsx
@@ -3,44 +3,6 @@ import styled from "styled-components";
 import "../../styles/TextStyle.css";
 import { useNavigate } from "react-router-dom";
 
-export const LoginOk = () => {
-  const navigate = useNavigate();
-
-  return (
-    <>
-      <Container>
-        <HomeLogo src={logo} alt="linkrap image" />
-        <div
-          style={{
-            display: "flex",
-            justifyContent: "center",
-            alignContent: "center",
-            alignItems: "center",
-          }}
-        >
-          <LoginBox>
-            <div
-              className="heading3"
-              style={{
-                textAlign: "center",
-                letterSpacing: 2,
-                lineHeight: "150%",
-              }}
-            >
-              집게 장전 완료!
-              <br />
-              이제 링크를 하나씩 담아보세요.
-            </div>
-            <LoginButton className="heading4" onClick={() => navigate("/home")}>
-              시작하기
-            </LoginButton>
-          </LoginBox>
-        </div>
-      </Container>
-    </>
-  );
-};
-
 const Container = styled.div`
   display: flex;
   flex-direction: row;
@@ -57,8 +19,8 @@ const HomeLogo = styled.img`
 // 주황색 박스
 const LoginBox = styled.div`
   border-radius: 100px;
-  width: 700px;
-  height: 800px;
+  width: 765px;
+  height: 906px;
   background-color: rgba(255, 160, 122, 0.3);
   display: flex;
   flex-direction: column;
@@ -74,9 +36,36 @@ const LoginButton = styled.button`
   width: 150;
   display: flex;
   justify-content: center;
-  color: #ffffff;
+  color: #2f2f2f;
   box-shadow: 0 4px 4px 0 rgba(0, 0, 0, 0.25);
   width: 154px;
   border-width: 0;
   cursor: pointer;
 `;
+
+export const LoginOk = () => {
+  const navigate = useNavigate();
+
+  return (
+    <>
+      <Container>
+        <HomeLogo src={logo} alt="linkrap image" />
+        <div
+          style={{
+            display: "flex",
+            justifyContent: "center",
+            alignContent: "center",
+            alignItems: "center",
+          }}
+        >
+          <LoginBox>
+            로그인 성공 페이지
+            <LoginButton className="heading3" onClick={() => navigate("/home")}>
+              Enter
+            </LoginButton>
+          </LoginBox>
+        </div>
+      </Container>
+    </>
+  );
+};
